# backend/openapi/openapi.v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Liveprofit API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/event-studies/review/prelabel": {
      "post": {
        "tags": [
          "event-studies"
        ],
        "summary": "Prelabel",
        "operationId": "prelabel_api_v1_event_studies_review_prelabel_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrelabelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_PrelabelData_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/event-studies/review/batch": {
      "post": {
        "tags": [
          "event-studies"
        ],
        "summary": "Submit Batch",
        "operationId": "submit_batch_api_v1_event_studies_review_batch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewBatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_ReviewBatchData_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/event-studies/review/events/{event_id}/compute": {
      "post": {
        "tags": [
          "event-studies"
        ],
        "summary": "Compute",
        "operationId": "compute_api_v1_event_studies_review_events__event_id__compute_post",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Event Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComputeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_ComputeData_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/event-studies/review/impact-drafts/{event_id}/confirm": {
      "post": {
        "tags": [
          "event-studies"
        ],
        "summary": "Confirm Impacts",
        "operationId": "confirm_impacts_api_v1_event_studies_review_impact_drafts__event_id__confirm_post",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Event Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmImpactsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_ConfirmImpactsData_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BatchSummary": {
        "properties": {
          "approved": {
            "type": "integer",
            "title": "Approved"
          },
          "ignored": {
            "type": "integer",
            "title": "Ignored"
          },
          "computed": {
            "type": "integer",
            "title": "Computed"
          }
        },
        "type": "object",
        "required": [
          "approved",
          "ignored",
          "computed"
        ],
        "title": "BatchSummary"
      },
...
      "ComputeData": {
        "properties": {
          "event_id": {
            "type": "integer",
            "title": "Event Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "ok",
              "failed"
            ],
            "title": "Status"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "event_id",
          "status"
        ],
        "title": "ComputeData"
      },
      "ComputeRequest": {
        "properties": {
          "operator": {
            "type": "string",
            "maxLength": 64,
            "title": "Operator",
            "default": "admin"
          }
        },
        "type": "object",
        "title": "ComputeRequest"
      },
...
      "ConfirmImpactsData": {
        "properties": {
          "event_id": {
            "type": "integer",
            "title": "Event Id"
          },
          "inserted": {
            "type": "integer",
            "title": "Inserted"
          }
        },
        "type": "object",
        "required": [
          "event_id",
          "inserted"
        ],
        "title": "ConfirmImpactsData"
      },
      "ConfirmImpactsRequest": {
        "properties": {
          "tickers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Tickers"
          },
          "operator": {
            "type": "string",
            "maxLength": 64,
            "title": "Operator",
            "default": "admin"
          }
        },
        "type": "object",
        "required": [
          "tickers"
        ],
        "title": "ConfirmImpactsRequest"
      },
...
      "EnvelopeMeta": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "description": "请求标识（X-Trace-ID）"
          },
          "schema_version": {
            "type": "string",
            "title": "Schema Version",
            "default": "v1"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor",
            "description": "不透明续页游标；无下一页为 null"
          }
        },
        "type": "object",
        "required": [
          "request_id"
        ],
        "title": "EnvelopeMeta"
      },
...
      "Envelope_ComputeData_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ComputeData"
          },
          "meta": {
            "$ref": "#/components/schemas/EnvelopeMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "Envelope[ComputeData]"
      },
...
      "Envelope_ConfirmImpactsData_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ConfirmImpactsData"
          },
          "meta": {
            "$ref": "#/components/schemas/EnvelopeMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "Envelope[ConfirmImpactsData]"
      },
...
      "Envelope_PrelabelData_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/PrelabelData"
          },
          "meta": {
            "$ref": "#/components/schemas/EnvelopeMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "Envelope[PrelabelData]"
      },
...
      "Envelope_ReviewBatchData_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ReviewBatchData"
          },
          "meta": {
            "$ref": "#/components/schemas/EnvelopeMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "Envelope[ReviewBatchData]"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PrelabelData": {
        "properties": {
          "prelabeled": {
            "type": "integer",
            "title": "Prelabeled"
          },
          "remaining": {
            "type": "integer",
            "title": "Remaining"
          }
        },
        "type": "object",
        "required": [
          "prelabeled",
          "remaining"
        ],
        "title": "PrelabelData"
      },
      "PrelabelRequest": {
        "properties": {
          "limit": {
            "type": "integer",
            "maximum": 200.0,
            "minimum": 1.0,
            "title": "Limit",
            "default": 50
          },
          "draft_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array",
                "maxItems": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Draft Ids"
          }
        },
        "type": "object",
        "title": "PrelabelRequest"
      },
...
      "ReviewBatchData": {
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/ReviewRowResult"
            },
            "type": "array",
            "title": "Results"
          },
          "summary": {
            "$ref": "#/components/schemas/BatchSummary"
          }
        },
        "type": "object",
        "required": [
          "results",
          "summary"
        ],
        "title": "ReviewBatchData"
      },
      "ReviewBatchRequest": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ReviewRowRequest"
            },
            "type": "array",
            "maxItems": 50,
            "minItems": 1,
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "ReviewBatchRequest"
      },
      "ReviewRowRequest": {
        "properties": {
          "draft_id": {
            "type": "integer",
            "title": "Draft Id"
          },
          "action": {
            "type": "string",
            "enum": [
              "approve",
              "ignore"
            ],
            "title": "Action"
          },
          "event_type": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 10000
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Type"
          },
          "event_subtype": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 10000
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Subtype"
          },
          "event_condition": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 10000
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Condition"
          },
          "importance": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 5.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Importance"
          },
          "expected_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Value"
          },
          "actual_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actual Value"
          },
          "previous_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous Value"
          },
          "event_scope": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 10000
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Scope"
          },
          "affected_scope_refs": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "maxLength": 10000
                },
                "type": "array",
                "maxItems": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Affected Scope Refs"
          },
          "operator": {
            "type": "string",
            "maxLength": 10000,
            "title": "Operator",
            "default": "admin"
          }
        },
        "type": "object",
        "required": [
          "draft_id",
          "action"
        ],
        "title": "ReviewRowRequest"
      },
      "ReviewRowResult": {
        "properties": {
          "draft_id": {
            "type": "integer",
            "title": "Draft Id"
          },
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "event_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Id"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "compute_status": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "ok",
                  "failed",
                  "skipped"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Compute Status"
          }
        },
        "type": "object",
        "required": [
          "draft_id",
          "ok"
        ],
        "title": "ReviewRowResult"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}